feat: error handling in get_parameter_value
No longer coercing these errors into None. Discovered amusing "feature" of the SDK types. The SSM client `get_parameter` call returns an error ParameterNotFound if the parameter is not found. If the parameter _is_ found and no other error is encountered, it returns a response with an Ok variant with a `parameter` field that holds an `Option<Parameter>` value. But obviously, the `None` case makes no sense since we would have already returned a ParameterNotFound error in the case that there were no such parameter

Signed-off-by: Ian Evans <icevans@icloud.com>